LayoutTests/inspector/audit/run-accessibility-expected.txt

Tests for the injected WebInspectorAudit.Accessibility functions.



== Running test suite: Audit.run.Accessibility
-- Running test case: Audit.run.Accessibility.getElementsByComputedRole.FakeRole
Audit setup...
Audit run `WebInspectorAudit.Accessibility.getElementsByComputedRole("FakeRole")`...
Result: []
Audit teardown...

-- Running test case: Audit.run.Accessibility.getElementsByComputedRole.FakeRole.parent
Audit setup...
Audit run `WebInspectorAudit.Accessibility.getElementsByComputedRole("FakeRole", document.querySelector("#parent"))`...
Result: []
Audit teardown...

-- Running test case: Audit.run.Accessibility.getElementsByComputedRole.tree
Audit setup...
Audit run `WebInspectorAudit.Accessibility.getElementsByComputedRole("tree")`...
Result: ["#parent"]
Audit teardown...

-- Running test case: Audit.run.Accessibility.getElementsByComputedRole.tree.parent
Audit setup...
Audit run `WebInspectorAudit.Accessibility.getElementsByComputedRole("tree", document.querySelector("#parent"))`...
Result: []
Audit teardown...

-- Running test case: Audit.run.Accessibility.getElementsByComputedRole.button
Audit setup...
Audit run `WebInspectorAudit.Accessibility.getElementsByComputedRole("button")`...
Result: ["#button","#link"]
Audit teardown...

-- Running test case: Audit.run.Accessibility.getElementsByComputedRole.button.parent
Audit setup...
Audit run `WebInspectorAudit.Accessibility.getElementsByComputedRole("button", document.querySelector("#parent"))`...
Result: []
Audit teardown...

-- Running test case: Audit.run.Accessibility.InvalidCopiedFunctionCall
Audit setup...
Copying WebInspectorAudit to window...
Audit teardown...
Testing copied getElementsByComputedRole...
PASS: Should produce an exception.
Error: NotAllowedError: Cannot be called outside of a Web Inspector Audit
